requirements.txt
# Aura Health — runtime dependencies
# Install: pip install -r requirements.txt

# LangGraph + LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-aws>=0.2.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0

# AWS
boto3>=1.34.0

# Vector store
faiss-cpu>=1.7.4
sentence-transformers>=2.6.0
huggingface-hub>=0.24.0

# PII scrubbing
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

# Jupyter
nest-asyncio>=1.6.0
ipywidgets>=8.0.0
jupyter>=1.0.0

# Utilities
python-dotenv>=1.0.0
